Code review - combine AxisAttributeMenu and LegendAttributeMenu into AxisOrLegendAttributeMenu - escape key closes menu - fixes bottom axis menu button placement - eliminates local management of menu open/close state in favor of letting Chakra handle it
